% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm4routing_extract.R
\name{nr_osm4routing}
\alias{nr_osm4routing}
\title{Extract Osmium Files Using osm4routing2}
\usage{
nr_osm4routing(
  osm_file = NULL,
  edges_file = "edges.csv",
  nodes_file = "nodes.csv",
  overwrite = FALSE
)
}
\arguments{
\item{osm_file}{osm.pbf file to be used}

\item{edges_file}{File name for the edges file.}

\item{nodes_file}{File name for the nodes file}

\item{overwrite}{Overwrite existing files (TRUE/FALSE)}
}
\value{
Node and edge files
}
\description{
Extract Osmium Files Using osm4routing2
}
\examples{
\dontrun{
# Without Compression
iow = osmextract::oe_get_network("isle of wight", mode = "driving")
?osmextract::oe_download
?osmextract::oe_match
region_name = "isle of wight"
pbf_info = osmextract::oe_match(region_name)
pbf_url = pbf_info$url
pbf_file = osmextract::oe_download(pbf_info$url)
nr_osm4routing(
  osm_file = pbf_file,
  edges_file = "edges.csv",
  nodes_file = "nodes.csv",
  overwrite = TRUE
)

# With Compression
nr_osm4routing(
  osm_file = pbf_file,
  edges_file = "edges.csv.gz",
  nodes_file = "nodes.csv.gz",
  overwrite = TRUE
)
}
}
