% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm4routing_extract.R
\name{osm4routing_extract}
\alias{osm4routing_extract}
\title{Extract Osmium Files Using osm4routing2}
\usage{
osm4routing_extract(
  osm_file = NULL,
  edges_file = "edges.csv",
  nodes_file = "nodes.csv",
  overwrite = FALSE
)
}
\arguments{
\item{osm_file}{osm.pbf file to be used}

\item{edges_file}{File name for the edges file.}

\item{nodes_file}{File name for the nodes file}

\item{overwrite}{Overwrite existing files (TRUE/FALSE)}
}
\value{
Node and edge files
}
\description{
Extract Osmium Files Using osm4routing2
}
\examples{
\dontrun{
# Without Compression
osm4routing_extract(osm_file = "example.osm.pbf",
'                   edges_file = "edges.csv",
'                   nodes_file = "nodes.csv",
'                   overwrite = TRUE)

# With Compression
osm4routing_extract(osm_file = "example.osm.pbf",
'                   edges_file = "edges.csv.gz",
'                   nodes_file = "nodes.csv.gz",
'                   overwrite = TRUE)
}
}
